+++ src/Device/Connect/connectCommand.ts
@@ -9,6 +9,18 @@ import { AppletDoesNotExistError } from "../../Applet/appletErrors";
 import RestApi from "@signageos/sdk/dist/RestApi/RestApi";
 import chalk from "chalk";
 
+const SERVER_PUBLIC_URL_OPTION = {
+	name: 'server-public-url',
+	type: String,
+	description: 'Public url of local machine server. Is useful when the local machine is behind a reverse proxy.',
+} as const;
+
+const SERVER_PORT_OPTION = {
+	name: 'server-port',
+	type: Number,
+	description: 'The custom server port for local machine server. Default is random.',
+} as const;
+
 const OPTION_LIST = [
 	NO_DEFAULT_ORGANIZATION_OPTION,
 	ORGANIZATION_UID_OPTION,
@@ -16,6 +28,8 @@ const OPTION_LIST = [
 	{ name: 'device-uid', type: String, description: 'Uid of device from box' },
 	APPLET_UID_OPTION,
 	APPLET_PATH_OPTION,
+	SERVER_PUBLIC_URL_OPTION,
+	SERVER_PORT_OPTION,
 ] as const;
 
 export const connect = createCommandDefinition({
@@ -34,8 +48,10 @@ export const connect = createCommandDefinition({
 		const deviceUid = await getDeviceUid(restApi, options);
 		const deviceData = await restApi.device.get(deviceUid);
 		await createConnectFile(deviceData.uid);
-		const serverData  = await serveApplet(projectDirAbsolutePath, appletUid, appletVersion, deviceData);
-		await connectDevice(organization, deviceData.uid, appletUid, appletVersion, serverData.serverPort);
+		const serverPublicUrl = options[SERVER_PUBLIC_URL_OPTION.name];
+		const serverPort = options[SERVER_PORT_OPTION.name];
+		const serverData  = await serveApplet(projectDirAbsolutePath, appletUid, appletVersion, deviceData, serverPort, serverPublicUrl);
+		await connectDevice(organization, deviceData.uid, appletUid, appletVersion, serverData.serverPublicUrl);
 
 		process.on('SIGINT', function () {
 			stopApplication(organization, deviceData.uid);