+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The `sos device connect` is looking for the applet in server by `name` instead of required `uid` in `package.json` `sos.appletUid` property
 - The `.git` folder is ignored automatically when the `.gitignore` file is used
 
+### Added
+- Allow customize `--server-port` and `--server-public-url` for `sos device connect` command
+
 ## [1.1.5] - 2023-01-02
 ### Fixed
 - Respect argument `--api-url` as priority over `SOS_API_URL` environment variable and default value